rebase completed, one failing test to fix
