+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/ImagenModel.kt
@@ -20,14 +20,10 @@ import com.google.firebase.appcheck.interop.InteropAppCheckTokenProvider
 import com.google.firebase.auth.internal.InternalAuthProvider
 import com.google.firebase.vertexai.common.APIController
 import com.google.firebase.vertexai.common.ContentBlockedException
-import com.google.firebase.vertexai.internal.GenerateImageRequest
-import com.google.firebase.vertexai.internal.GenerateImageResponse
-import com.google.firebase.vertexai.internal.ImagenParameters
-import com.google.firebase.vertexai.internal.ImagenPromptInstance
-import com.google.firebase.vertexai.internal.util.AppCheckHeaderProvider
-import com.google.firebase.vertexai.internal.util.toInternal
-import com.google.firebase.vertexai.internal.util.toPublicGCS
-import com.google.firebase.vertexai.internal.util.toPublicInline
+import com.google.firebase.vertexai.common.AppCheckHeaderProvider
+import com.google.firebase.vertexai.common.GenerateImageRequest
+import com.google.firebase.vertexai.common.ImagenParameters
+import com.google.firebase.vertexai.common.ImagenPromptInstance
 import com.google.firebase.vertexai.type.FirebaseVertexAIException
 import com.google.firebase.vertexai.type.ImagenGCSImage
 import com.google.firebase.vertexai.type.ImagenGenerationConfig
@@ -133,7 +129,7 @@ internal constructor(
   }
 }
 
-private fun GenerateImageResponse.validate(): GenerateImageResponse {
+private fun ImagenGenerationResponse.Internal.validate(): ImagenGenerationResponse.Internal {
   if (predictions.none { it.mimeType != null }) {
     throw ContentBlockedException(
       message = predictions.first { it.raiFilteredReason != null }.raiFilteredReason
