--- firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/internal/GenerateImageResponse.kt
@@ -1,30 +1,0 @@
-/*
- * Copyright 2025 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.google.firebase.vertexai.internal
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class GenerateImageResponse(val predictions: List<ImagenImageResponse>) {}
-
-@Serializable
-internal data class ImagenImageResponse(
-  val bytesBase64Encoded: String? = null,
-  val gcsUri: String? = null,
-  val mimeType: String? = null,
-  val raiFilteredReason: String? = null,
-)
