+++ firebase-vertexai/src/main/kotlin/com/google/firebase/vertexai/type/ImagenImageFormat.kt
@@ -16,22 +16,29 @@
 
 package com.google.firebase.vertexai.type
 
+import kotlinx.serialization.Serializable
+
 /**
  * Represents the format an image should be returned in.
+ *
  * @param mimeType A string (like `"image/jpeg"`) specifying the encoding MIME type of the image.
  * @param compressionQuality an int (1-100) representing the quality of the image; a lower number
- * means the image is permitted to be lower quality to reduce size. This parameter is not relevant
- * for every MIME type.
+ *   means the image is permitted to be lower quality to reduce size. This parameter is not relevant
+ *   for every MIME type.
  */
 public class ImagenImageFormat
 private constructor(public val mimeType: String, public val compressionQuality: Int?) {
+
+  internal fun toInternal() = Internal(mimeType, compressionQuality)
+
+  @Serializable internal data class Internal(val mimeType: String, val compressionQuality: Int?)
 
   public companion object {
     /**
      * An [ImagenImageFormat] representing a JPEG image.
      *
      * @param compressionQuality an int (1-100) representing the quality of the image; a lower
-     * number means the image is permitted to be lower quality to reduce size.
+     *   number means the image is permitted to be lower quality to reduce size.
      */
     public fun jpeg(compressionQuality: Int? = null): ImagenImageFormat {
       return ImagenImageFormat("image/jpeg", compressionQuality)
